requirements.txt
django
gunicorn
django-heroku
requests
django-rest-hooks
django-rest-framework
pypdftk
python-dotenv
